TestDiscovery error handling and related refactoring.
Add support for handling two types of errors during test discovery:
1. JUnit or AJUR detects an invalid or missing test class
2. A process crash occurs, such as an exception during the application creation flow

The first type of error was handled by converting TestDiscovery to a RunListener, so it
has access to the failure information reported by JUnit.
The second type was handled by adding additional logic to onException.

TestEventClient was also refactored to attempt to encapsulate the details of
its service connection.

PiperOrigin-RevId: 407502083